+++ src/main/java/application/domain/Product.java
@@ -43,13 +43,13 @@ public class Product {
     @Getter
 
     @Setter
-    private long reductionPercentage;
+    private Double reductionPercentage;
     @Getter
     @Setter
-    private long previousPrice;
+    private Double previousPrice;
     @Getter
     @Setter
-    private long currentPrice;
+    private Double currentPrice;
     @Getter
     @Setter
     private Category category;
